chore: moved CSP to next.config. rm middleware
